nano carves: runtime TUs shed ast.h - standalone_ctx_utils, gc pod split, aot registration re-home, CodeOfPolicies header move; explicit options stack honored exactly
