+++ include/daScript/simulate/standalone_ctx_utils.h
@@ -2,13 +2,17 @@
 #define DAS_AST_SIM_CONVERTERS_H
 
 #include <daScript/misc/platform.h>
+#include <daScript/misc/anyhash.h>
 #include <daScript/simulate/simulate.h>
-#include <daScript/ast/ast.h>
-
-#include <optional>
+#include <daScript/simulate/aot_library.h>
+#include <daScript/simulate/code_of_policies.h>
 
 namespace das {
     using MangledNameHash = uint64_t;
+
+    inline MangledNameHash mangledNameHashOf ( const string & mangledName ) {
+        return hash_blockz64((uint8_t *)mangledName.c_str());
+    }
 
     struct FunctionInfo {
         FunctionInfo() = delete;
@@ -43,7 +47,7 @@ namespace das {
         GlobalVarInfo() = delete;
         GlobalVarInfo(string name, const string &mangledName, uint32_t typeSize, bool globalShared)
             : name(das::move(name))
-            , mangledNameHash(Variable::getMNHash(mangledName))
+            , mangledNameHash(mangledNameHashOf(mangledName))
             , typeSize(typeSize)
             , globalShared(globalShared) {}
