+++ tests/aot/_standalone_zero_stack_fixture.das
@@ -1,0 +1,7 @@
+options gen2
+options stack = 0
+
+[export]
+def zero_stack_probe : int {
+    return 1
+}
